backend/pyproject.toml
[project]
name = "missionout-backend"
version = "0.1.0"
description = "FastAPI backend for MissionOut SAR alerting"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "alembic>=1.14.0",
  "pydantic-settings>=2.6.1",
  "google-auth>=2.35.0",
  "requests>=2.32.0",
  "pyjwt[crypto]>=2.9.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5",
  "httpx>=0.27",
  "jsonschema>=4.21",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
